# Ensure unbind action is always available on request