@smoogipoo 2022-09-16
With my commit above:

Before:
|                  Method |          Mean |      Error |     StdDev |  Gen 0 | Allocated |
|------------------------ |--------------:|-----------:|-----------:|-------:|----------:|
|            TestBaseline |     0.0000 ns |  0.0000 ns |  0.0000 ns |      - |         - |
|   TestCompositeDrawable | 2,718.5841 ns | 16.9034 ns | 15.8115 ns | 0.0191 |   1,744 B |
|           TestContainer | 3,020.8970 ns | 33.6598 ns | 31.4854 ns | 0.0191 |   1,816 B |
| TestTypeNestedComposite | 3,111.6984 ns | 30.8866 ns | 27.3802 ns | 0.0191 |   1,744 B |

After:
|                  Method |          Mean |      Error |     StdDev |  Gen 0 | Allocated |
|------------------------ |--------------:|-----------:|-----------:|-------:|----------:|
|            TestBaseline |     0.0000 ns |  0.0000 ns |  0.0000 ns |      - |         - |
|   TestCompositeDrawable | 2,684.0064 ns | 11.7135 ns | 10.3837 ns | 0.0191 |   1,680 B |
|           TestContainer | 2,950.6035 ns | 14.5887 ns | 12.9325 ns | 0.0191 |   1,752 B |
| TestTypeNestedComposite | 3,166.6835 ns | 22.7095 ns | 21.2425 ns | 0.0191 |   1,680 B |

I honestly don't know whether optimising this has been worth it at all, given the number of issues and time investment in fixing those issues as a result. So if this doesn't work then I propose we revert to the known "working" version.