@peppy 2022-09-16
The original optimisation was due to a considerable overhead in actual game runtime, so I would argue it is worth it.